Rewrite ll_read_block and ll_write_block
Read and write subsections of objects rather than reading an entire
object and discarding the excess or reading an object, modifying a
subsection, and then writing the whole thing.

+++ src/include/cephfs/libcephfs.h
@@ -323,10 +323,10 @@ int ceph_ll_osdaddr(struct ceph_mount_info *cmount,
 		    int osd, uint32_t *addr);
 uint64_t ceph_ll_get_internal_offset(struct ceph_mount_info *cmount,
 				     struct vinodeno_t vino, uint64_t blockno);
-uint64_t ceph_ll_read_block(struct ceph_mount_info *cmount,
-			    struct vinodeno_t vino, uint64_t blockid,
-			    char* bl, uint64_t offset, uint64_t length,
-			    struct ceph_file_layout* layout);
+int ceph_ll_read_block(struct ceph_mount_info *cmount,
+		       struct vinodeno_t vino, uint64_t blockid,
+		       char* bl, uint64_t offset, uint64_t length,
+		       struct ceph_file_layout* layout);
 int ceph_ll_write_block(struct ceph_mount_info *cmount, struct vinodeno_t vino,
 			uint64_t blockid, char* buf, uint64_t offset,
 			uint64_t length, struct ceph_file_layout* layout,
